+++ changelog.md
@@ -176,6 +176,7 @@ its results in a DBF file that is read by the MLogit.R script. "RJDBC" is indeed
     - Warn the user if the number of threads set for an assignment is larger than the number of physical cores.
     - Tested with Java 11
     - The assignments are now canceled (with an error message) if a path has a non strictly positive cost.
+    - SystemInfoDlg replaces ArchInfoDlg and gives now also information about the system hardware.
 
 - Breaking changes:
     - Simplify the API for modal split methods. This breaks existing plugin's. If an incompatible plugin is found, an error